Day 19 part 1 and refactoring of Program class

+++ src/main/kotlin/mpdev/springboot/aoc2018/solutions/day16/Day16.kt
@@ -38,7 +38,7 @@ class Day16: PuzzleSolver() {
         val elapsed = measureTimeMillis {
             programUtils.identifyOpCodes()
             programUtils.executeProgram()
-            result = programUtils.getRegister()[0]
+            result = programUtils.getRegister()[0].toInt()
         }
         return PuzzlePartSolution(2, result.toString(), elapsed)
     }
